fix: PR #1 코드 리뷰 피드백 반영하여 Critical Issues 해결
## 해결된 문제점들:

### 🔴 Critical Issues
- **auth-helpers.ts**: window.location.origin 사용 함수들을 별도 파일로 분리하고 browser client 사용
  - lib/auth-helpers-client.ts 생성 (resetPassword, signInWithGoogle, signInWithGithub)
  - 서버 전용 함수들은 기존 파일에 유지
- **document-helpers.ts**: undefined supabase 참조를 (await createClient()) 사용으로 수정
- **share-helpers.ts**: incomplete migration 완료, 모든 supabase 참조 수정
- **storage-helpers.ts**: createServiceRoleClient 참조를 createClient로 수정

### 🟡 Important Issues
- **unused variables**: auth-helpers.ts에서 사용되지 않는 변수 선언들 제거
- **build verification**: 모든 수정사항이 빌드를 통과하는지 확인

이제 모든 helper 파일들이 올바른 클라이언트를 사용하고 runtime error 없이 동작합니다.
