+++ lib/auth-helpers-client.ts
@@ -1,0 +1,60 @@
+import { createClient } from './supabase/client'
+import type { AuthError } from '@supabase/supabase-js'
+
+// Browser-only auth functions that use window.location.origin
+export async function resetPassword(email: string) {
+  try {
+    const supabase = createClient()
+    const { error } = await supabase.auth.resetPasswordForEmail(email, {
+      redirectTo: `${window.location.origin}/reset-password`
+    })
+
+    if (error) {
+      throw error
+    }
+
+    return { error: null }
+  } catch (error) {
+    return { error: error as AuthError }
+  }
+}
+
+export async function signInWithGoogle() {
+  try {
+    const supabase = createClient()
+    const { data, error } = await supabase.auth.signInWithOAuth({
+      provider: 'google',
+      options: {
+        redirectTo: `${window.location.origin}/auth/callback`
+      }
+    })
+
+    if (error) {
+      throw error
+    }
+
+    return { data, error: null }
+  } catch (error) {
+    return { data: null, error: error as AuthError }
+  }
+}
+
+export async function signInWithGithub() {
+  try {
+    const supabase = createClient()
+    const { data, error } = await supabase.auth.signInWithOAuth({
+      provider: 'github',
+      options: {
+        redirectTo: `${window.location.origin}/auth/callback`
+      }
+    })
+
+    if (error) {
+      throw error
+    }
+
+    return { data, error: null }
+  } catch (error) {
+    return { data: null, error: error as AuthError }
+  }
+}
